requirements.txt
django==1.11.1
gunicorn==19.6.0
dj-database-url==0.4.1
mysqlclient==1.3.12
whitenoise==3.2
django-dotenv == 1.4.2
django-crispy-forms==1.7.2